// src/SharedModule/Services/Models/UserStatsDto.ts
export class UserStatsDto {
    Id: number;
    ContactId: string;
    SessionId: string;
    Url: string;
    Browser: string;
    Ip: string;
    Area: string;
    Controller: string;
    Action: string;
    Details: string;
    CreatedDate: string;
    CreatedId: number;
    LasteditedDate: string;
    LasteditedId: number;
}